Abstract the ENGDB_Service
